refactor(api): Bot相关内容包路径调整
