+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/bot/Bot.kt
@@ -14,12 +14,13 @@
  *
  */
 
-package love.forte.simbot
+package love.forte.simbot.bot
 
 import kotlinx.coroutines.CompletionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancel
+import love.forte.simbot.*
 import love.forte.simbot.ability.Survivable
 import love.forte.simbot.definition.User
 import love.forte.simbot.definition.UserInfo
