+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/bot/BotManager.kt
@@ -14,8 +14,9 @@
  *
  */
 
-package love.forte.simbot
+package love.forte.simbot.bot
 
+import love.forte.simbot.*
 import love.forte.simbot.application.EventProvider
