+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/definition/Member.kt
@@ -18,7 +18,7 @@ package love.forte.simbot.definition
 
 import kotlinx.coroutines.flow.firstOrNull
 import love.forte.simbot.Api4J
-import love.forte.simbot.Bot
+import love.forte.simbot.bot.Bot
 import love.forte.simbot.ID
 import love.forte.simbot.Timestamp
 import love.forte.simbot.action.MuteSupport
