+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/definition/Objective.kt
@@ -17,7 +17,7 @@
 package love.forte.simbot.definition
 
 import love.forte.simbot.Api4J
-import love.forte.simbot.Bot
+import love.forte.simbot.bot.Bot
 import love.forte.simbot.ID
 import love.forte.simbot.action.SendSupport
 import love.forte.simbot.action.sendIfSupport
