+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/definition/Stranger.kt
@@ -16,7 +16,7 @@
 
 package love.forte.simbot.definition
 
-import love.forte.simbot.Bot
+import love.forte.simbot.bot.Bot
 import love.forte.simbot.ID
