+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/definition/User.kt
@@ -18,7 +18,7 @@
 
 package love.forte.simbot.definition
 
-import love.forte.simbot.Bot
+import love.forte.simbot.bot.Bot
 import love.forte.simbot.ID
 
 /**
