+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/event/MessageEvents.kt
@@ -17,7 +17,7 @@
 package love.forte.simbot.event
 
 import love.forte.simbot.Api4J
-import love.forte.simbot.Bot
+import love.forte.simbot.bot.Bot
 import love.forte.simbot.ID
 import love.forte.simbot.action.ReplySupport
 import love.forte.simbot.action.SendSupport
