+++ simbot-apis/simbot-api/src/main/kotlin/love/forte/simbot/event/RequestEvents.kt
@@ -18,7 +18,7 @@ package love.forte.simbot.event
 
 import kotlinx.coroutines.launch
 import love.forte.simbot.Api4J
-import love.forte.simbot.Bot
+import love.forte.simbot.bot.Bot
 import love.forte.simbot.ExperimentalSimbotApi
 import love.forte.simbot.ID
 import love.forte.simbot.definition.*
