+++ simbot-cores/simbot-core/src/main/kotlin/love/forte/simbot/core/application/BaseApplicationBuilder.kt
@@ -21,6 +21,10 @@ import love.forte.simbot.*
 import love.forte.simbot.ability.CompletionPerceivable
 import love.forte.simbot.application.*
 import love.forte.simbot.application.BotRegistrar
+import love.forte.simbot.bot.Bot
+import love.forte.simbot.bot.BotManager
+import love.forte.simbot.bot.BotVerifyInfo
+import love.forte.simbot.bot.ComponentMismatchException
 import love.forte.simbot.event.EventProcessor
 import love.forte.simbot.utils.view
 import java.util.concurrent.ConcurrentLinkedQueue
@@ -210,7 +214,7 @@ public abstract class BaseApplicationBuilder<A : Application> :
         override fun register(botVerifyInfo: BotVerifyInfo): Bot? {
             logger.info("Registering bot with verify info [{}]", botVerifyInfo)
             for (manager in providers) {
-                if (manager !is love.forte.simbot.BotRegistrar) {
+                if (manager !is love.forte.simbot.bot.BotRegistrar) {
                     continue
                 }
